requirements.txt
Flask>=3.0.3
google-cloud-storage>=2.18.2
pyrebase4>=4.8.0
firebase-admin>=6.5.0
python-dotenv>=1.0.1